refactor: 이미지 수정 로직 바텀시트로 분리.
