+++ src/pages/set-user-info/SetUserInfo.tsx
@@ -1,9 +1,8 @@
-import { DEFAULT_PROFILE_IMAGE } from '@pages/mypage/hooks/use-user-data';
 import Button from '@shared/components/button/Button';
 import { Icon } from '@shared/components/icon/Icon';
 import Loading from '@shared/components/loading/Loading';
 import Navigation from '@shared/components/navigation/Navigation';
-import React, { useState } from 'react';
+import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import SetUserName from '@pages/set-user-info/components/SetUserName';
 import SetUserEmail from '@pages/set-user-info/components/SetUserEmail';
@@ -11,15 +10,12 @@ import SetUserGender from '@pages/set-user-info/components/SetUserGender';
 import SetAgreement from '@pages/set-user-info/components/SetAgreement';
 import SetUserImage from '@pages/set-user-info/components/SetUserImage';
 import ProfileImageBottomSheet from '@pages/set-user-info/components/ProfileImageBottomSheet';
-import useToast from '@shared/hooks/use-toast';
-import { isAcceptableFile, isFileSizeValid } from '@shared/utils/image';
-import { MAX_MB, NOT_ALLOWED_FILE_TYPE } from '@shared/constant/image';
 import { FormProvider } from 'react-hook-form';
 import { useSetUserInfo } from '@pages/set-user-info/hooks/use-set-user-info';
 
 export default function SetUserInfo() {
   const navigate = useNavigate();
-  const toast = useToast();
+
   const [isBottomSheetOpen, setIsBottomSheetOpen] = useState(false);
 
   const { formMethods, handleSubmit, isFetching, isValid, userData } =
@@ -33,32 +29,6 @@ export default function SetUserInfo() {
   };
   const handleClickBack = () => navigate(-1);
 
-  const handleFileChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files?.[0];
-    if (!file) return;
-
-    if (!isAcceptableFile(file)) {
-      toast.error(NOT_ALLOWED_FILE_TYPE);
-      return;
-    }
-
-    if (!isFileSizeValid(file)) {
-      toast.error(`파일 용량은 ${MAX_MB}MB 이하여야 합니다.`);
-      return;
-    }
-
-    // TODO : 추후 presignedURL api로 대체
-    const imageUrl = URL.createObjectURL(file);
-
-    formMethods.setValue('profileImageUrl', imageUrl);
-    handleCloseBottomSheet();
-  };
-
-  const handleDeleteImage = () => {
-    formMethods.setValue('profileImageUrl', DEFAULT_PROFILE_IMAGE);
-    handleCloseBottomSheet();
-  };
-
   if (isFetching || !userData) {
     return <Loading />;
   }
@@ -69,8 +39,6 @@ export default function SetUserInfo() {
         <ProfileImageBottomSheet
           isBottomSheetOpen={isBottomSheetOpen}
           handleCloseBottomSheet={handleCloseBottomSheet}
-          handleFileChange={handleFileChange}
-          handleDeleteImage={handleDeleteImage}
         />
         <Navigation
           text={'프로필 수정'}
